refactor stash

+++ src/engine/Common.h
@@ -14,7 +14,7 @@
 
 namespace staywalk
 {
-	using idtype = int64_t;
+	using idtype = unsigned long long;
 
 	using vec2 = glm::vec2;
 	using vec3 = glm::vec3;
